// admin-clients/libs/cpanel/promoters/events/sessions/data-access/src/lib/models/session-release-status.enum.ts
export enum SessionReleaseStatus {
    planned = 'PLANNED',
    inProgramming = 'IN_PROGRAMMING',
    releasePending = 'RELEASE_PENDING',
    released = 'RELEASED',
    releaseCancelled = 'RELEASE_CANCELLED',
    releaseFinished = 'RELEASE_FINISHED',
    cancelled = 'CANCELLED',
    notAccomplished = 'NOT_ACCOMPLISHED'
}