// admin-clients/libs/cpanel/promoters/events/sessions/data-access/src/lib/models/session-sale-status.enum.ts
export enum SessionSaleStatus {
    planned = 'PLANNED',
    inProgramming = 'IN_PROGRAMMING',
    salePending = 'SALE_PENDING',
    saleCancelled = 'SALE_CANCELLED',
    pendingSaleCancelled = 'PENDING_SALE_CANCELLED',
    sale = 'SALE',
    saleFinished = 'SALE_FINISHED',
    cancelled = 'CANCELLED',
    notAccomplished = 'NOT_ACCOMPLISHED'
}